AVR EEPROM is only compiled on AVR fixed printf warnings on the host (Linux 64-bit)

+++ src/frogfs_assert.h
@@ -26,8 +26,8 @@
 #define FROGFS_ASSERT(x,y,...)             do { uint32_t line = __LINE__;           \
                                                   if (x != y)                        \
                                                   {                                  \
-                                                      FROGFS_PRINTF(FROGFS_DEBUG_STR_MEM("assertion failed at line %lu: "), line); \
-                                                      FROGFS_PRINTF(FROGFS_DEBUG_STR_MEM("was %08lx, expected %08lx"), (uint32_t)x, (uint32_t)y);  \
+                                                      FROGFS_PRINTF(FROGFS_DEBUG_STR_MEM("assertion failed at line %lu: "), (unsigned long)line); \
+                                                      FROGFS_PRINTF(FROGFS_DEBUG_STR_MEM("was %08lx, expected %08lx"), (unsigned long)x, (unsigned long)y);  \
                                                       FROGFS_PRINTF(FROGFS_DEBUG_STR_MEM("\r\n"));                  \
                                                       exit(1);                       \
                                                   }                                  \
@@ -36,16 +36,16 @@
 #define FROGFS_ASSERT_VERBOSE(x,y,fmt,...)      do { uint32_t line = __LINE__;          \
                                                   if (x != y)                        \
                                                   {                                  \
-                                                      FROGFS_PRINTF(FROGFS_DEBUG_STR_MEM("assertion failed at line %lu: "), line); \
+                                                      FROGFS_PRINTF(FROGFS_DEBUG_STR_MEM("assertion failed at line %lu: "), (unsigned long)line); \
                                                       FROGFS_PRINTF(FROGFS_DEBUG_STR_MEM(fmt), ## __VA_ARGS__);           \
-                                                      FROGFS_PRINTF(FROGFS_DEBUG_STR_MEM("was %08lx, expected %08lx"), (uint32_t)x, (uint32_t)y);  \
+                                                      FROGFS_PRINTF(FROGFS_DEBUG_STR_MEM("was %08lx, expected %08lx"), (unsigned long)x, (unsigned long)y);  \
                                                       FROGFS_PRINTF(FROGFS_DEBUG_STR_MEM("\r\n"));                  \
                                                       exit(1);                       \
                                                   }                                  \
                                              }while(0);
 
 #define FROGFS_ASSERT_UNCHECKED(fmt,...)      do { uint32_t line = __LINE__;          \
-                                                  FROGFS_PRINTF(FROGFS_DEBUG_STR_MEM("assertion failed at line %lu: "), line); \
+                                                  FROGFS_PRINTF(FROGFS_DEBUG_STR_MEM("assertion failed at line %lu: "), (unsigned long)line); \
                                                   FROGFS_PRINTF(FROGFS_DEBUG_STR_MEM(fmt), ## __VA_ARGS__);           \
                                                   FROGFS_PRINTF(FROGFS_DEBUG_STR_MEM("\r\n"));                  \
                                                   exit(1);                       \
